requirements.txt
Flask
tensorflow
numpy
pillow
Flask-CORS